Move the Filter logic into a custom hook.
